Extracting ImguiRttFontCache to its own file
